weak_supervised_cross_modal/requirements.txt
torch>=1.12.0
torchvision>=0.13.0
transformers>=4.21.0
numpy>=1.21.0
scipy>=1.7.0
scikit-learn>=1.0.0
matplotlib>=3.5.0
seaborn>=0.11.0
tensorboard>=2.8.0
wandb>=0.12.0
einops>=0.4.0
timm>=0.6.0
opencv-python>=4.5.0
Pillow>=8.3.0
tqdm>=4.62.0
pyyaml>=6.0
dataclasses>=0.6
typing-extensions>=4.0.0
psutil>=5.8.0
torchvision>=0.13.0
Pillow>=8.3.0 